Initial Commit - Completed Project
This is the initial commit of the completed code file. light_sensor_test.c  is the main code file and was written completely by me. ece198.c and ece198.h define the fundamental pin and timer initialization functions that light_sensor_test.c uses and were written by Bernie Roehl. They are referenced in light_sensor_test.c with his permission but are not added to this repository as of now.